Put ICEkitContentsAdmin and ICEkitFluentContentsAdmin back into icekit.admin for nicer overall structure and separation of icekit.admin_tools.mixins concerns
