+++ icekit/admin_mixins.py
@@ -1,18 +1,21 @@
 import warnings
 
-warnings.warn(
-    "the icekit.admin_mixins module is deprecated. Use icekit.admin_tools.mixins instead.",
-    DeprecationWarning,
-    stacklevel=2)
-
-
 from icekit.utils.deprecation import deprecated
 
 from icekit.admin_tools.mixins import \
     FluentLayoutsMixin as new_FluentLayoutsMixin, \
     HeroMixinAdmin as new_HeroMixinAdmin, \
-    ListableMixinAdmin as new_ListableMixinAdmin, \
+    ListableMixinAdmin as new_ListableMixinAdmin
+
+from icekit.admin_tools.polymorphic import \
     ICEkitFluentPagesParentAdmin as new_ICEkitFluentPagesParentAdmin
+
+
+warnings.warn(
+    "the icekit.admin_mixins module is deprecated. Use icekit.admin_tools.mixins instead.",
+    DeprecationWarning,
+    stacklevel=2)
+
 
 @deprecated
 class FluentLayoutsMixin(new_FluentLayoutsMixin):
@@ -22,6 +25,7 @@ class FluentLayoutsMixin(new_FluentLayoutsMixin):
     """
     pass
 
+
 @deprecated
 class HeroMixinAdmin(new_HeroMixinAdmin):
     """
@@ -30,6 +34,7 @@ class HeroMixinAdmin(new_HeroMixinAdmin):
     """
     pass
 
+
 @deprecated
 class ListableMixinAdmin(new_ListableMixinAdmin):
     """
@@ -37,6 +42,7 @@ class ListableMixinAdmin(new_ListableMixinAdmin):
     Use :class:`icekit.admin_tools.mixins.ListableMixinAdmin` instead.
     """
     pass
+
 
 @deprecated
 class ICEkitFluentPagesParentAdmin(new_ICEkitFluentPagesParentAdmin):
