+++ icekit/admin_tools/mixins.py
@@ -72,62 +72,6 @@ class ListableMixinAdmin(admin.ModelAdmin):
     )
 
 
-# This import must go here to avoid class loading errors
-from icekit.publishing.admin import ICEKitFluentPagesParentAdminMixin
-
-
-# WARNING: Beware of very closely named classes here
-class ICEkitFluentPagesParentAdmin(
-        ICEKitFluentPagesParentAdminMixin, WorkflowMixinAdmin):
-    """
-    A base for Fluent Pages parent admins that will include ICEkit features:
-
-     - publishing
-     - workflow
-    """
-    # Go through contortions here to make sure the 'actions_column' is last
-    list_display = [
-        display_column for display_column in (
-            ICEKitFluentPagesParentAdminMixin.list_display +
-            WorkflowMixinAdmin.list_display)
-        if display_column != 'actions_column'] + ['actions_column']
-    list_filter = ICEKitFluentPagesParentAdminMixin.list_filter + \
-        WorkflowMixinAdmin.list_filter
-    inlines = [WorkflowStateTabularInline]
-
-
-# import has to happen here to avoid circular import errors
-from icekit.publishing.admin import PublishingAdmin, \
-    PublishableFluentContentsAdmin
-
-
-class ICEkitContentsAdmin(PublishingAdmin, WorkflowMixinAdmin):
-    """
-    A base for generic admins that will include ICEkit features:
-
-     - publishing
-     - workflow
-    """
-    list_display = PublishingAdmin.list_display + \
-        WorkflowMixinAdmin.list_display
-    list_filter = PublishingAdmin.list_filter + \
-        WorkflowMixinAdmin.list_filter
-    inlines = [WorkflowStateTabularInline]
-
-
-class ICEkitFluentContentsAdmin(
-        PublishableFluentContentsAdmin, WorkflowMixinAdmin):
-    """
-    A base for Fluent Contents admins that will include ICEkit features:
-
-     - publishing
-     - workflow
-    """
-    list_display = ICEkitContentsAdmin.list_display
-    list_filter = ICEkitContentsAdmin.list_filter
-    inlines = ICEkitContentsAdmin.inlines
-
-
 class ThumbnailAdminMixin(object):
     """
     Shortcut for displaying a thumbnail in a changelist (or inline).
